app/requirements.txt
fastapi
uvicorn
neo4j
sentence-transformers
pymongo
numpy
scikit-learn
websockets
python-multipart
aiofiles
pillow